requirements.txt
langchain
langchain-community
langchain-groq
langchain-pinecone
langchain-huggingface
langchain_core
python-dotenv